+++ src/modes/interOrderbook.ts
@@ -2,7 +2,7 @@ import { orderbookAbi } from "../abis";
 import { estimateGasCost } from "../gas";
 import { BigNumber, Contract, ethers } from "ethers";
 import { containsNodeError, errorSnapshot } from "../error";
-import { getBountyEnsureRainlang, parseRainlang } from "../config";
+import { getBountyEnsureRainlang, parseRainlang } from "../task";
 import { BaseError, ExecutionRevertedError, PublicClient } from "viem";
 import { ONE18, scale18To, estimateProfit, withBigintSerializer } from "../utils";
 import { BotConfig, BundledOrders, ViemClient, DryrunResult, SpanAttrs } from "../types";
