+++ src/modes/intraOrderbook.ts
@@ -2,7 +2,7 @@ import { orderbookAbi } from "../abis";
 import { estimateGasCost } from "../gas";
 import { BigNumber, ethers } from "ethers";
 import { containsNodeError, errorSnapshot } from "../error";
-import { getWithdrawEnsureRainlang, parseRainlang } from "../config";
+import { getWithdrawEnsureRainlang, parseRainlang } from "../task";
 import { estimateProfit, scale18, withBigintSerializer } from "../utils";
 import { BaseError, erc20Abi, ExecutionRevertedError, PublicClient } from "viem";
 import {
